refactor: telemetry package consolidation
