+++ internal/cmd/apit/setsnippet.go
@@ -10,7 +10,6 @@ import (
 	"github.com/saucelabs/saucectl/internal/apitest"
 	cmds "github.com/saucelabs/saucectl/internal/cmd"
 	"github.com/saucelabs/saucectl/internal/http"
-	"github.com/saucelabs/saucectl/internal/segment"
 	"github.com/saucelabs/saucectl/internal/usage"
 	"github.com/spf13/cobra"
 )
@@ -44,7 +43,7 @@ cat snippet2.xml | saucectl apit vault set-snippet snip2 - --project "smoke test
 				return fmt.Errorf("invalid HTTP_PROXY value")
 			}
 
-			tracker := segment.DefaultClient
+			tracker := usage.DefaultClient
 
 			go func() {
 				tracker.Collect(
