+++ internal/cmd/ini/cypress.go
@@ -11,7 +11,6 @@ import (
 	"github.com/saucelabs/saucectl/internal/config"
 	"github.com/saucelabs/saucectl/internal/cypress"
 	v1 "github.com/saucelabs/saucectl/internal/cypress/v1"
-	"github.com/saucelabs/saucectl/internal/segment"
 	"github.com/saucelabs/saucectl/internal/usage"
 	"github.com/spf13/cobra"
 )
@@ -26,7 +25,7 @@ func CypressCmd() *cobra.Command {
 		Short:        "Bootstrap a Cypress project.",
 		SilenceUsage: true,
 		Run: func(cmd *cobra.Command, _ []string) {
-			tracker := segment.DefaultClient
+			tracker := usage.DefaultClient
 
 			go func() {
 				tracker.Collect(
