+++ internal/cmd/storage/cmd.go
@@ -4,10 +4,11 @@ import (
 	"errors"
 	"time"
 
+	"github.com/saucelabs/saucectl/internal/usage"
+
 	"github.com/saucelabs/saucectl/internal/credentials"
 	"github.com/saucelabs/saucectl/internal/http"
 	"github.com/saucelabs/saucectl/internal/region"
-	"github.com/saucelabs/saucectl/internal/segment"
 	"github.com/spf13/cobra"
 )
 
@@ -33,7 +34,7 @@ func Command(preRun func(cmd *cobra.Command, args []string)) *cobra.Command {
 				return errors.New("invalid region")
 			}
 			if reg == region.Staging {
-				segment.DefaultClient.Enabled = false
+				usage.DefaultClient.Enabled = false
 			}
 
 			appsClient = *http.NewAppStore(reg.APIBaseURL(),
