+++ internal/cmd/storage/delete.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	cmds "github.com/saucelabs/saucectl/internal/cmd"
-	"github.com/saucelabs/saucectl/internal/segment"
 	"github.com/saucelabs/saucectl/internal/usage"
 	"github.com/spf13/cobra"
 )
@@ -23,7 +22,7 @@ func DeleteCommand() *cobra.Command {
 			return nil
 		},
 		PreRun: func(cmd *cobra.Command, _ []string) {
-			tracker := segment.DefaultClient
+			tracker := usage.DefaultClient
 
 			go func() {
 				tracker.Collect(
